phase2(safety): extract editRiskScore.scoreEditFromContext as a pure tested core
The edit risk/confidence score gates auto-apply (HIGH can never be silently
auto-approved; YOLO keys its silent-apply threshold off the 0.2/0.7 boundaries), but
EditRiskScoringService.scoreEdit was untested. Extract everything except factor #6
(the count of pre-existing Error markers, which needs the live model/markers) into
pure common/editRiskScore.ts scoreEditFromContext(context, existingErrorCount). The
service now computes only that count -- passing 0 when the model is unavailable or
there's no newContent, mirroring the old `if (model && newContent)` guard -- and
delegates. EditContext/EditRiskScore types moved to the pure module and re-exported
from the service so existing import paths are unchanged. Byte-identical logic.

Tested: factor-by-factor golden cases (deletion=HIGH+1.0, critical +0.5, >50% rewrite
crosses HIGH, test-file +0.2, multi-file cap, >5-errors +0.2, create floor 0.05, tiny
edit very-low) + the LOW/MEDIUM/HIGH classifier boundaries incl. the silent-auto-apply
boundary (riskScore<0.2 AND confidence>0.7). tsgo 0; 685 tests.
